Move iree_time_now() impl into base/internal.
