+++ runtime/src/iree/base/internal/CMakeLists.txt
@@ -332,6 +332,17 @@ iree_cc_test(
     iree::testing::gtest
     iree::testing::gtest_main
 )
+iree_cc_library(
+  NAME
+    time
+  HDRS
+    "time.h"
+  SRCS
+    "time.c"
+  DEPS
+    iree::base::core_headers
+  PUBLIC
+)
 
 iree_cc_library(
   NAME
