+++ runtime/src/iree/base/tracing/BUILD.bazel
@@ -68,6 +68,7 @@ iree_runtime_cc_library(
     ],
     deps = [
         "//runtime/src/iree/base:core_headers",
+        "//runtime/src/iree/base/internal:time",
     ],
 )
